% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_size}
\alias{check_size}
\title{check_size}
\usage{
check_size(input)
}
\arguments{
\item{input}{object to check}
}
\description{
Checks if input Seurat object is small enough to run clustOpt
}
\keyword{internal}
