% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_sil_plots}
\alias{create_sil_plots}
\title{create_sil_plots}
\usage{
create_sil_plots(sil_dist)
}
\arguments{
\item{sil_dist}{the output of clustOpt}
}
\value{
A list of ggplot objects
}
\description{
Silhouette score distribution plots
}
