% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{prep_train}
\alias{prep_train}
\title{prep_train}
\usage{
prep_train(
  input,
  subject_ids,
  dtype = "scRNA",
  within_batch = NA,
  test_id,
  verbose = 0
)
}
\arguments{
\item{input}{Seurat object}

\item{subject_ids}{Metadata field that identifies unique samples.}

\item{dtype}{Type of data in the Seurat object "scRNA" or "CyTOF", default
is "scRNA". CyTOF data is expected to be arcsinh normalized.}

\item{within_batch}{Batch variable, for a given sample only those with the
same value for the batch variable will be used for training.}

\item{test_id}{subject_id for the test sample}

\item{verbose}{Integer verbosity level (0 = silent, 1 = milestones,
2 = detailed, 3 = includes Seurat output)}
}
\value{
Training data formatted for sil_score, format depends on dtype
}
\description{
Prepare training data for random forest
}
