% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustOpt.R
\name{prepare_CyTOF}
\alias{prepare_CyTOF}
\title{prepare_CyTOF}
\usage{
prepare_CyTOF(train_seurat, test_seurat)
}
\arguments{
\item{train_seurat}{A Seurat object representing the training data set.}

\item{test_seurat}{A Seurat object representing the test data set.}
}
\value{
A list of the training and test data formatted for RF
}
\description{
Prepare CyTOF data for random forest
}
