% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustOpt.R
\name{calculate_mse_score}
\alias{calculate_mse_score}
\title{calculate_mse_score}
\usage{
calculate_mse_score(predicted, data_frame)
}
\arguments{
\item{predicted}{Cluster assignments}

\item{data_frame}{Data frame containing the data}
}
\value{
A numeric value representing the mse
}
\description{
Calculate mean squared error score
}
