% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustOpt.R
\name{calculate_silhouette_score}
\alias{calculate_silhouette_score}
\title{Calculate silhouette score}
\usage{
calculate_silhouette_score(predicted, data_frame)
}
\arguments{
\item{predicted}{Cluster assignments}

\item{data_frame}{Data frame containing the data}
}
\value{
A list containing the average silhouette score and the average
silhouette score for each cluster.
}
\description{
Calculate silhouette score
}
