% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_size}
\alias{check_size}
\title{check_size: Checks if input Seurat object is small enough to run
sil_score}
\usage{
check_size(input, verbose = TRUE)
}
\arguments{
\item{input}{object to check}

\item{verbose}{print warning about size}
}
\description{
check_size: Checks if input Seurat object is small enough to run
sil_score
}
