% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustOpt.R
\name{clust_opt}
\alias{clust_opt}
\title{clust_opt}
\usage{
clust_opt(
  input,
  ndim,
  dtype = "scRNA",
  sketch_size = NULL,
  skip_sketch = FALSE,
  subject_ids,
  res_range = c(0.02, 0.04, 0.06, 0.08, 0.1, 0.2, 0.4, 0.6, 0.8, 1, 1.2),
  within_batch = NA,
  verbose = FALSE,
  num_trees = 1000,
  split_method = "odd_even",
  train_with = "A",
  min_cells = 50
)
}
\arguments{
\item{input}{Seurat object}

\item{ndim}{Number of principal components to use.}

\item{dtype}{Type of data in the Seurat object "scRNA" or "CyTOF", default
is "scRNA". CyTOF data is expected to be arcsinh normalized  (in the counts
slot) and sketching is not implemented for CyTOF.}

\item{sketch_size}{Number of cells to use for sketching.}

\item{skip_sketch}{Skip sketching, by default any input with more than
200,000 cells is sketched to 10\% of the cells.}

\item{subject_ids}{Metadata field that identifies unique samples.}

\item{res_range}{Range of resolutions to test.}

\item{within_batch}{Batch variable, for a given sample only those with the
same value for the batch variable will be used for training.}

\item{verbose}{Output messages.}

\item{num_trees}{Number of trees to use in the random forest.}

\item{split_method}{Which method of PCA splitting (for train and test) should
be used "odd_even" or "var_balanced", default is "odd_even".}

\item{train_with}{Either A or B, for split_method = "odd_even" A is odd and
B is even. Default is A}

\item{min_cells}{Minimum cells per subject, default is 50}
}
\value{
A data.frame containing a distribution of silhouette scores for each
resolution.
}
\description{
Runs the main resolution optimization algorithm
}
