% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustOpt.R
\name{clust_opt}
\alias{clust_opt}
\title{clust_opt: Runs the main resolution optimization algorithm}
\usage{
clust_opt(
  input,
  ndim,
  normalization = "SCT",
  sketch_size = NULL,
  subject_ids,
  seed = 1,
  res_range = c(0.02, 0.04, 0.06, 0.08, 0.1, 0.2, 0.4, 0.6, 0.8, 1, 1.2),
  ncore = 12,
  verbose = TRUE
)
}
\arguments{
\item{input}{Seurat object}

\item{ndim}{Number of principal components to use.}

\item{normalization}{Method to use for normalizing counts}

\item{subject_ids}{Metadata field that identifies unique samples.}

\item{res_range}{Range of resolutions to test.}

\item{ncore}{Number of cores, should match the number of resolutions being
tested.}

\item{verbose}{print messages.}

\item{within_batch}{Batch variable, for a given sample only those with the 
same value for the batch variable will be used for training.}
}
\value{
A data.frame containing a distribution of silhouette scores for each
resolution.
}
\description{
clust_opt: Runs the main resolution optimization algorithm
}
