% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{leverage_sketch}
\alias{leverage_sketch}
\title{leverage_sketch}
\usage{
leverage_sketch(input, sketch_size, verbose = TRUE)
}
\arguments{
\item{input}{Seurat object}

\item{sketch_size}{Number of cells to include in the sketch assay}

\item{verbose}{print messages}
}
\value{
Seurat object with sketch assay
}
\description{
Uses leverage score based sampling to reduce the size of the
input Seurat object and create a sketch assay using this subsample
}
