% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustOpt.R
\name{prepare_cytof}
\alias{prepare_cytof}
\title{prepare_cytof}
\usage{
prepare_cytof(train_seurat, test_seurat)
}
\arguments{
\item{train_seurat}{A Seurat object representing the training data set.}

\item{test_seurat}{A Seurat object representing the test data set.}
}
\value{
A list of the training and test data formatted for RF
}
\description{
Prepare CyTOF data for random forest
}
