% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustOpt.R
\name{project_pca}
\alias{project_pca}
\title{project_pca}
\usage{
project_pca(train_seurat, test_seurat, train_with, verbose)
}
\arguments{
\item{train_seurat}{A Seurat object representing the training data set.}

\item{test_seurat}{A Seurat object representing the test data set.}

\item{train_with}{Which set A or B should be used for training}

\item{verbose}{print messages}
}
\value{
A list containing two elements: `train_df` and `test_df`. Each is a matrix
of the projected data for the training and test sets, respectively, using
the specified number of principal components.
}
\description{
Project Training and Test Seurat Objects onto Principal Components

This function projects both training and test Seurat objects onto a set of
principal components derived from the training data.
}
\details{
Identifies features that are common between the training
and test data sets. Extracts the PCA loadings from the training data
for common features. If train_with = "A", then projection is performed
with only the B PCs. Both training and test data are projected onto these
loadings. Data is also projected onto loadings from all PCs for cluster
scoring after training is complete.
}
