% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustOpt.R
\name{project_PCA}
\alias{project_PCA}
\title{Project Training and Test Seurat Objects onto Principal Components}
\usage{
project_PCA(train_seurat, test_seurat, ndim)
}
\arguments{
\item{train_seurat}{A Seurat object representing the training data set.}

\item{test_seurat}{A Seurat object representing the test data set.}

\item{ndim}{The number of principal components to use for projection.}
}
\value{
A list containing two elements: `train_df` and `test_df`. Each is a matrix
of the projected data for the training and test sets, respectively, using
the specified number of principal components.
}
\description{
This function projects both training and test Seurat objects onto a set of
principal components derived from the training data.
}
\details{
Identifies features that are common between the training
and test data sets. It then extracts the PCA loadings from the training data
for these common features. Both training and test data are projected onto
these loadings. Finally, the function selects the specified number of
dimensions (principal components) for the output.
}
