% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sil_summary}
\alias{sil_summary}
\title{sil_summary}
\usage{
sil_summary(input)
}
\arguments{
\item{input}{output of clustOpt}
}
\value{
A data.frame summarising the silhouette score distribution
}
\description{
Summarises the silhouette score distribution output by clustOpt
}
