% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{split_pca_dimensions}
\alias{split_pca_dimensions}
\title{split_pca_dimensions}
\usage{
split_pca_dimensions(input, verbose = 0)
}
\arguments{
\item{input}{Seurat object}

\item{verbose}{Integer verbosity level (0 = silent, 1 = milestones,
2 = detailed, 3 = includes Seurat output, 4 = includes output from other
packages such as ranger)}
}
\value{
Seurat object with new PCA reductions
}
\description{
Takes a Seurat object with an existing PCA reduction and splits the
dimensions into 2 sets odd and even PCs.
}
\keyword{internal}
