% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_split.R
\name{split_pca_dimensions}
\alias{split_pca_dimensions}
\title{split_pca_dimensions}
\usage{
split_pca_dimensions(input, verbose = FALSE)
}
\arguments{
\item{input}{Seurat object}

\item{verbose}{Output messages}
}
\value{
Seurat object with new PCA reductions
}
\description{
Takes a Seurat object with an existing PCA reduction and splits the
dimensions into 2 sets odd and even PCs.
}
