% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_split.R
\name{split_pca_dimensions}
\alias{split_pca_dimensions}
\title{split_pca_dimensions}
\usage{
split_pca_dimensions(input, split_method = "odd_even", verbose = FALSE)
}
\arguments{
\item{input}{Seurat object}

\item{split_method}{How should the PCs be split into 2, options are odd even
split ("odd_even") or balancing variation explained between 2 sets
("var_balanced").}

\item{verbose}{Ouput messages}
}
\value{
Seurat object with new PCA reductions
}
\description{
Takes a Seurat object with an existing PCA reduction and splits the
dimensions into 2 sets A and B. The split_method is either "odd_even" or
"var_balanced". The last PC will be removed if there are an odd number of
PCs. The new reductions will be named "A_pca" and "B_pca" with A being odd
and B being even for split_method = "odd_even".
}
