% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustOpt.R
\name{split_pca_dimensions}
\alias{split_pca_dimensions}
\title{split_pca_dimensions}
\usage{
split_pca_dimensions(input, balance = TRUE)
}
\arguments{
\item{input}{Seurat object}

\item{balance}{Balance the number of even and odd PC dimensions, default
is TRUE}
}
\value{
Seurat object with new PCA reductions
}
\description{
Takes a Seurat object with an existing PCA reduction and splits the
dimensions into even and odd PC dimensions. New reductions are
created and named "even_pca" and "odd_pca". The last PC will be
removed if there are an odd number of PCs.
}
\examples{
\dontrun{
seurat_obj <- CreateSeuratObject(counts = matrix(rnorm(1000), ncol = 10))
seurat_obj <- RunPCA(seurat_obj)
seurat_obj <- split_pca_by_parity(seurat_obj)
}
}
