% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustOpt.R
\name{train_random_forest}
\alias{train_random_forest}
\title{train_random_forest}
\usage{
train_random_forest(res, df_list, train_clusters, sam, num.trees)
}
\arguments{
\item{res}{Resolution to train on}

\item{df_list}{List containing training and test data}

\item{train_clusters}{Cluster assignments for the training data}

\item{sam}{Test sample}
}
\value{
A list containing the resolution, silhouette score, and number of
predicted clusters.
}
\description{
Train the random forest and predict on the test sample
}
